Change User IDs to long, change all SparseArrays to LongSparseArrays

+++ SmartMap/src/ch/epfl/smartmap/cache/Event.java
@@ -39,7 +39,7 @@ public interface Event {
     /**
      * @return The ID of the user who created the event
      */
-    int getCreator();
+    long getCreator();
     
     /**
      * @return The event's ID
